% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSolar.R
\name{eot}
\alias{eot}
\title{Equation of time}
\usage{
eot(DOY)
}
\arguments{
\item{DOY}{Day of year}
}
\description{
This function approximate the value of equation of time for a given date.
}
\examples{
eot()
}
\keyword{Equation}
\keyword{of}
\keyword{time}
\keyword{value}
