% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LST.R
\name{LST}
\alias{LST}
\title{Local Standard Time}
\usage{
LST(DOY)
}
\arguments{
\item{DOY}{Day of year}
}
\description{
This function returns local standard time (in minutes) given a day of the year value.
}
\examples{

#Calculating  local standard time for two consecutive days

DOY <- seq(0, 2, .05)

lst <- LST(DOY)

plot(DOY, lst)

}
\keyword{LST}
