% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sunset.R
\name{Sunset}
\alias{Sunset}
\title{Sunset Time}
\usage{
Sunset(DOY, Lat)
}
\arguments{
\item{DOY}{Day of year}

\item{Lat}{Latitude (in degrees)}
}
\description{
This function estimates sunset time (in continuous hour values) for a given day of year and latitude.
}
\examples{

#Calculating sunset time for 365 day of the year for 45 degree latitude

DOY <- 1:365

Lat = 45

sunset <- Sunset(DOY, Lat)

plot(DOY, sunset)

}
\keyword{Sunset}
